Resolved build warnings

+++ OpenGL/OpenGL/src/VertexArray.cpp
@@ -15,7 +15,7 @@ void VertexArray::AddBuffer(const VertexBuffer& vb, const VertexBufferLayout& la
 {
 	vb.Bind();
 
-	unsigned int offset = 0;
+	uintptr_t offset = 0;
 
 	const auto& elements = layout.GetElements();
 	for (unsigned int i = 0; i < elements.size(); i++)
